move TaskModel and SmartSurveyTask down to backbone
